YeongChingZhou/jc_requirements.txt
# Core PyTorch and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision and Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.9.0

# Data Analysis and Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0

# Machine Learning Metrics and Utilities
scikit-learn>=1.1.0

# Progress Bars and CLI
tqdm>=4.64.0

# Additional utilities (if needed)
imageio>=2.19.0
albumentations>=1.3.0

